+++ lib/testcentricity_web/exception_queue_helper.rb
@@ -100,7 +100,7 @@ module TestCentricity
         Capybara.save_screenshot "#{path}.png"
       end
       puts "Screenshot saved at #{path}.png"
-      screen_shot = { :path => path, :filename => filename }
+      screen_shot = {path: path, filename: filename}
       Environ.save_screen_shot(screen_shot)
     end
   end