+++ lib/testcentricity_web/web_elements/select_list.rb
@@ -3,14 +3,16 @@ module TestCentricity
     attr_accessor :list_item
     attr_accessor :selected_item
     attr_accessor :list_trigger
+    attr_accessor :text_field
 
     def initialize(name, parent, locator, context)
       super
       @type = :selectlist
       list_spec = {
-          :list_item     => "li[class*='active-result']",
-          :selected_item => "li[class*='result-selected']",
-          :list_trigger  => nil
+          list_item:     "li[class*='active-result']",
+          selected_item: "li[class*='result-selected']",
+          list_trigger:  nil,
+          text_field:    nil
       }
       define_list_elements(list_spec)
     end
@@ -18,14 +20,16 @@ module TestCentricity
     def define_list_elements(element_spec)
       element_spec.each do |element, value|
         case element
-          when :list_item
-            @list_item = value
-          when :selected_item
-            @selected_item = value
-          when :list_trigger
-            @list_trigger = value
-          else
-            raise "#{element} is not a recognized selectlist element"
+        when :list_item
+          @list_item = value
+        when :selected_item
+          @selected_item = value
+        when :list_trigger
+          @list_trigger = value
+        when :text_field
+          @text_field = value
+        else
+          raise "#{element} is not a recognized selectlist element"
         end
       end
     end
@@ -39,9 +43,9 @@ module TestCentricity
     #
     # @example
     #   province_select.choose_option('Alberta')
-    #   province_select.choose_option(:value => 'AB')
-    #   state_select.choose_option(:index => 24)
-    #   state_select.choose_option(:text => 'Maryland')
+    #   province_select.choose_option(value: 'AB')
+    #   state_select.choose_option(index: 24)
+    #   state_select.choose_option(text: 'Maryland')
     #
     def choose_option(option)
       obj, = find_element
@@ -59,9 +63,9 @@ module TestCentricity
           end
         else
           if option.is_a?(Hash)
-            page.find(:css, "#{@list_item}:nth-of-type(#{option[:index]})").click if option.has_key?(:index)
-            page.find(:css, "#{@list_item}:nth-of-type(#{option[:value]})").click if option.has_key?(:value)
-            page.find(:css, "#{@list_item}:nth-of-type(#{option[:text]})").click if option.has_key?(:text)
+            page.find(:css, "#{@list_item}:nth-of-type(#{option[:index]})").click if option.key?(:index)
+            page.find(:css, "#{@list_item}:nth-of-type(#{option[:value]})").click if option.key?(:value)
+            page.find(:css, "#{@list_item}:nth-of-type(#{option[:text]})").click if option.key?(:text)
           else
             options = obj.all(@list_item).collect(&:text)
             sleep(2) unless options.include?(option)
@@ -79,6 +83,20 @@ module TestCentricity
       end
     end
 
+    def set(text)
+      if @text_field.nil?
+        raise "A 'text_field' list element must be defined before calling the 'set' method on a selectlist object"
+      end
+      obj, = find_element
+      object_not_found_exception(obj, nil)
+      if @list_trigger.nil?
+        obj.click
+      else
+        page.find(:css, @list_trigger).click
+      end
+      page.find(:css, @text_field, wait: 2).set(text)
+    end
+
     # Return array of strings of all options in a select box object.
     # Supports standard HTML select objects and Chosen select objects.
     #
@@ -89,10 +107,10 @@ module TestCentricity
     def get_options
       obj, = find_element
       object_not_found_exception(obj, nil)
-      if first(:css, @list_item, minimum: 0)
+      if obj.first(:css, @list_item, minimum: 0)
         obj.all(@list_item).collect(&:text)
       else
-        obj.all('option').collect(&:text)
+        obj.all('option', visible: :all).collect(&:text)
       end
     end
 
@@ -108,10 +126,10 @@ module TestCentricity
     def get_option_count
       obj, = find_element
       object_not_found_exception(obj, nil)
-      if first(:css, @list_item, minimum: 0)
+      if obj.first(:css, @list_item, minimum: 0)
         obj.all(@list_item).count
       else
-        obj.all('option').count
+        obj.all('option', visible: :all).count
       end
     end
 
@@ -119,9 +137,11 @@ module TestCentricity
 
     def verify_options(expected, enqueue = false)
       actual = get_options
-      enqueue ?
-          ExceptionQueue.enqueue_assert_equal(expected, actual, "Expected list of options in list #{object_ref_message}") :
-          assert_equal(expected, actual, "Expected list of options in list #{object_ref_message} to be #{expected} but found #{actual}")
+      if enqueue
+        ExceptionQueue.enqueue_assert_equal(expected, actual, "Expected list of options in list #{object_ref_message}")
+      else
+        assert_equal(expected, actual, "Expected list of options in list #{object_ref_message} to be #{expected} but found #{actual}")
+      end
     end
 
     # Return text of first selected option in a select box object.
@@ -134,10 +154,10 @@ module TestCentricity
     def get_selected_option
       obj, = find_element
       object_not_found_exception(obj, nil)
-      if first(:css, @list_item, minimum: 0)
+      if obj.first(:css, @list_item, minimum: 0)
         obj.first(:css, @selected_item).text
       else
-        obj.first('option[selected]').text
+        obj.first('option[selected]', visible: :all).text
       end
     end
 
@@ -152,7 +172,7 @@ module TestCentricity
     def choose_siebel_option(option)
       Capybara.wait_on_first_by_default = true
       invoke_siebel_popup
-      first(:xpath, "//li[@class='ui-menu-item']", :exact => true, :match => :prefer_exact, text: option).click
+      first(:xpath, "//li[@class='ui-menu-item']", exact: true, match: :prefer_exact, text: option).click
     end
 
     # Return array of strings of all options in a Siebel OUI select box object.
@@ -174,9 +194,11 @@ module TestCentricity
       invoke_siebel_popup
       sleep(0.5)
       actual = page.all(:xpath, "//li[@class='ui-menu-item']").collect(&:text)
-      enqueue ?
-          ExceptionQueue.enqueue_assert_equal(expected, actual, "Expected list of options in list #{object_ref_message}") :
-          assert_equal(expected, actual, "Expected list of options in list #{object_ref_message} to be #{expected} but found #{actual}")
+      if enqueue
+        ExceptionQueue.enqueue_assert_equal(expected, actual, "Expected list of options in list #{object_ref_message}")
+      else
+        assert_equal(expected, actual, "Expected list of options in list #{object_ref_message} to be #{expected} but found #{actual}")
+      end
       obj, = find_element
       obj.native.send_keys(:escape)
     end
@@ -197,9 +219,9 @@ module TestCentricity
 
     def select_item(obj, option)
       if option.is_a?(Hash)
-        obj.find("option[value='#{option[:value]}']").click if option.has_key?(:value)
+        obj.find("option[value='#{option[:value]}']").click if option.key?(:value)
 
-        if option.has_key?(:index)
+        if option.key?(:index)
           if @locator_type == :xpath
             obj.find(:xpath, "option[#{option[:index]}]").select_option
           else
@@ -207,9 +229,9 @@ module TestCentricity
           end
         end
 
-        obj.select option[:text] if option.has_key?(:text)
+        obj.select option[:text] if option.key?(:text)
       else
-        obj.select option
+        obj.select(option, visible: :all)
       end
     end
   end